week6_AI/requirements.txt
google-generativeai>=0.5.0
pydantic>=2.0.0
pymongo>=4.0.0
streamlit>=1.35.0
pandas>=2.0.0